refactor: moving the automation client into the common package
